pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "b2b_casino_backend"
version = "1.0.0"
description = "B2B Online Casino Integration Platform Backend"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi==0.108.0",
    "uvicorn[standard]==0.25.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "pyjwt==2.8.0",
    "cryptography==41.0.5",
    "redis==5.0.1",
    "fakeredis==2.20.0",
    "kafka-python==2.0.2",
    "aiokafka==0.8.1",
    "pydantic==2.5.2",
    "python-multipart==0.0.6",
    "email-validator==2.1.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
    "loguru==0.7.2",
    "psutil==5.9.6",
    "prometheus-client==0.19.0",
    "aiohttp",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "respx==0.20.2",
    "assertpy==1.1",
]

[project.urls]
"Homepage" = "https://github.com/your_username/your_repo"
"Bug Tracker" = "https://github.com/your_username/your_repo/issues"

[tool.setuptools.packages.find]
# where = ["src"]

[tool.pytest.ini_options]
pythonpath = [".", "backend"]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "api: API tests",
    "e2e: End-to-end tests",
    "slow: Mark tests that are slow and might be skipped in CI",
    "performance: marks tests as performance tests",
]
addopts = "--verbose --cov=backend --cov-report=term-missing --cov-report=html:coverage_html"
